% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{as_dataframe}
\alias{as_dataframe}
\title{Convert, matrix, list, data.frame, sf POINT or sfc_POINT object to data frame}
\usage{
as_dataframe(x, rename_xy = c("longitude", "latitude"), row_names = "index")
}
\arguments{
\item{x}{matrix, list, data.frame, sf POINT or sfc_POINT type object}

\item{rename_xy}{character, rename X and Y coordinates to these?  NULL to skip renaming.}

\item{row_names}{character, if "index" set row names to indices, "none" to skip}
}
\value{
data frame with [X,Y] or [longitude, latitude] as the leading columns
}
\description{
Convert, matrix, list, data.frame, sf POINT or sfc_POINT object to data frame
}
