% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{boyce.cm}
\alias{boyce.cm}
\title{Continuous Boyce Index (modified from ecospat package)}
\usage{
boyce.cm(fit, obs, nclass = 0, window.w = "default", res = 100, PEplot = TRUE)
}
\arguments{
\item{fit}{A vector or Raster-Layer containing the predicted suitability values}

\item{obs}{A vector containing the predicted suitability values or xy-coordinates (if "fit" is a Raster-Layer) of the validation points (presence records)}

\item{nclass}{The number of classes or vector with class thresholds. If nclass=0, the Boyce index is calculated with a moving window (see next parameters)}

\item{window.w}{The width of the moving window (by default 1/10 of the suitability range)}

\item{res}{The resolution of the moving window (by default 100 focals)}

\item{PEplot}{If true, plot the predicted to expected ratio along the suitability class}
}
\description{
This function was modified from ecospat::ecospat.boyce() by Cory
Merow to make it run faster. It retains the same functionality as the original
function. Function parameter information is taken from ecospat.boyce().
}
