% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmm.R
\name{buildRMM}
\alias{buildRMM}
\title{Build metadata object from ENMeval results}
\usage{
buildRMM(e, envs, rmm = NULL)
}
\arguments{
\item{e}{ENMevaluation object}

\item{envs}{RasterStack of environmental predictor variables used in analysis}
}
\description{
Builds a \code{rangeModelMetadata} object from the output of \code{ENMevaluate}.
See Merow et al. (2019) for more details on the nature of the metadata and the \code{rangeModelMetadata} package.
To improve reproducibility of the study, this metadata object can be used as supplemental information for a manuscript, shared with collaborators, etc.
}
