% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvariegatus.R
\docType{data}
\name{bvariegatus}
\alias{bvariegatus}
\title{Example occurrence dataset.}
\format{
A data frame with 476 rows and 2 variables:
\describe{
  \item{longitude}{The longitude coordinate (x).}
  \item{latitude}{The latitude coordinate (y).}
}
}
\source{
\url{https://www.gbif.org/}
}
\usage{
bvariegatus
}
\description{
An example occurrence dataset for the Brown Throated Sloth (\emph{Bradypus variegatus}) downloaded from GBIF with the \code{spocc} package.
}
\keyword{datasets}
