% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{clamp}
\alias{clamp}
\title{Clamp predictor variables}
\usage{
clamp(predictors, p.z, left = NULL, right = NULL, categoricals = NULL)
}
\arguments{
\item{predictors}{Raster* object; predictor variables in raster form (either Layer or Stack)}

\item{p.z}{matrix or data frame; the predictor variables values for the reference records
(not including coordinates), used to determine the minimums and maximums -- 
this should ideally be the occurrences + background (can be made with raster::extract())}

\item{left}{character; names of variables to get a minimum clamp (i.e. left))}

\item{right}{character; names of variables to get a maximum clamp (i.e. right))}

\item{categoricals}{character; name or names of categorical environmental variables}
}
\description{
Clamp predictor variables
}
\author{
Stephen J. Phillips, Jamie M. Kass, Gonzalo Pinilla-Buitrago
}
