% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{emp.bg}
\alias{emp.bg}
\alias{emp.bg,ENMnull-method}
\title{emp.bg generic for ENMnull object}
\usage{
emp.bg(x)

\S4method{emp.bg}{ENMnull}(x)
}
\arguments{
\item{x}{ENMnull object}
}
\description{
emp.bg generic for ENMnull object
}
