% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{emp.occs.grp}
\alias{emp.occs.grp}
\alias{emp.occs.grp,ENMnull-method}
\title{emp.occs.grp generic for ENMnull object}
\usage{
emp.occs.grp(x)

\S4method{emp.occs.grp}{ENMnull}(x)
}
\arguments{
\item{x}{ENMnull object}
}
\description{
emp.occs.grp generic for ENMnull object
}
