% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{emp.occs}
\alias{emp.occs}
\alias{emp.occs,ENMnull-method}
\title{emp.occs generic for ENMnull object}
\usage{
emp.occs(x)

\S4method{emp.occs}{ENMnull}(x)
}
\arguments{
\item{x}{ENMnull object}
}
\description{
emp.occs generic for ENMnull object
}
