% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{enm.args}
\alias{enm.args}
\alias{enm.args<-}
\alias{enm.args,ENMdetails-method}
\alias{enm.args<-,ENMdetails-method}
\title{enm.args generic for ENMdetails object}
\usage{
enm.args(x)

enm.args(x) <- value

\S4method{enm.args}{ENMdetails}(x)

\S4method{enm.args}{ENMdetails}(x) <- value
}
\arguments{
\item{x}{ENMdetails object}

\item{value}{input value}
}
\description{
enm.args generic for ENMdetails object
}
