% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm.bioclim.R
\docType{data}
\name{enm.bioclim}
\alias{enm.bioclim}
\title{ENMdetails bioclim}
\format{
An object of class \code{ENMdetails} of length 1.
}
\usage{
enm.bioclim
}
\description{
This is the ENMdetails implementation for the BIOCLIM climate envelope model, implemented by predicts.
}
\keyword{datasets}
