% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm.brt.R
\docType{data}
\name{enm.boostedRegressionTrees}
\alias{enm.boostedRegressionTrees}
\title{ENMdetails: Boosted Regression Trees}
\format{
An object of class \code{ENMdetails} of length 1.
}
\usage{
enm.boostedRegressionTrees
}
\description{
Boosted regression trees are implemented with the function \code{gbm.step} from the \code{dismo}
package. Details on this algorithm sand its uses in ecology and evolution can be found in Elith et al. (2008).
The function \code{gbm.step} runs a cross-validation procedure to determine the optimal number of trees by minimizing
holdout deviance (see ?dismo::gbm.step for details). As the number of trees (\code{n.trees}) is decided by \code{gbm.step} 
for each combination of the tunable parameters, \code{n.trees} is not itself tunable in ENMeval.\cr
\cr
The tunable parameters in this implementation are:\cr
1. \code{tree.complexity}: the number of variable interactions to allow (a value of 1 means no interactions, 
or an additive model); analogous to \code{interaction.depth} in the \code{gbm} package\cr
2. \code{learning.rate}: the contribution of each tree to the model (a smaller value increases the number of 
trees required, and smaller is generally preferable); analogous to \code{shrinkage} in the \code{gbm} package\cr

In order to run \code{gbm.step} with custom partitions, the partitions chosen by the user in ENMeval are specified 
with argument \code{fold.vector}. See Details for internal changes to partitions that are made for random k-fold 
so the function can run. In addition, \code{site.weights} are included for background records to balance the occurrence 
and background data during model fitting, equal to the number of occurrence records divided by the number of background records.
Besides these specifications, the function \code{gbm.step} is run with all defaults

Jackknife (leave-one-out) partitioning is not available for boosted regression trees mainly because the models often fail
with small datasets, and even if they do succeed to completion these models in general are not advisable for data-poor species
(pers. comm. Jane Elith).
}
\details{
As \code{gbm.step} requires that each class is partitioned for cross validation, both occurrence and background
records must have associated folds for the optimization procedure. Thus, for random k-fold partitioning, background records
are internally partitioned for \code{gbm.step}. In addition, validation models are built internally by \code{gbm.step}, and 
therefore are not built sequentially within \code{tune.enm.R} as for other algorithms -- this results in a implementation that
avoids running the optimization procedure multiple times.
}
\references{
Elith, J., Leathwick, J. R., & Hastie, T. (2008). A working guide to boosted regression trees. \emph{Journal of Animal Ecology}, 77(4), 802-813.
}
\keyword{datasets}
