% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{enm.errors}
\alias{enm.errors}
\alias{enm.errors<-}
\alias{enm.errors,ENMdetails-method}
\alias{enm.errors<-,ENMdetails-method}
\title{enm.errors generic for ENMdetails object}
\usage{
enm.errors(x)

enm.errors(x) <- value

\S4method{enm.errors}{ENMdetails}(x)

\S4method{enm.errors}{ENMdetails}(x) <- value
}
\arguments{
\item{x}{ENMdetails object}

\item{value}{input value}
}
\description{
enm.errors generic for ENMdetails object
}
