% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{enm.fun}
\alias{enm.fun}
\alias{enm.fun<-}
\alias{enm.fun,ENMdetails-method}
\alias{enm.fun<-,ENMdetails-method}
\title{enm.fun generic for ENMdetails object}
\usage{
enm.fun(x)

enm.fun(x) <- value

\S4method{enm.fun}{ENMdetails}(x)

\S4method{enm.fun}{ENMdetails}(x) <- value
}
\arguments{
\item{x}{ENMdetails object}

\item{value}{input value}
}
\description{
enm.fun generic for ENMdetails object
}
