% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm.gbm.R
\docType{data}
\name{enm.gbm}
\alias{enm.gbm}
\title{ENMdetails gbm}
\format{
An object of class \code{ENMdetails} of length 1.
}
\usage{
enm.gbm
}
\description{
This is the ENMdetails implementation for gradient boosting 
trees, implemented with the gbm package.
}
\keyword{datasets}
