% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{enm.msgs}
\alias{enm.msgs}
\alias{enm.msgs<-}
\alias{enm.msgs,ENMdetails-method}
\alias{enm.msgs<-,ENMdetails-method}
\title{enm.msgs generic for ENMdetails object}
\usage{
enm.msgs(x)

enm.msgs(x) <- value

\S4method{enm.msgs}{ENMdetails}(x)

\S4method{enm.msgs}{ENMdetails}(x) <- value
}
\arguments{
\item{x}{ENMdetails object}

\item{value}{input value}
}
\description{
enm.msgs generic for ENMdetails object
}
