% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{enm.name}
\alias{enm.name}
\title{ENMdetails <- function(name, fun, errors, msgs, args, predict, ncoefs, varimp) {
  new("ENMdetails", name = name, fun = fun, errors = errors, msgs = msgs, 
      args = args, predict = predict, ncoefs = ncoefs, varimp = varimp)
}}
\usage{
enm.name(x)
}
\description{
ENMdetails <- function(name, fun, errors, msgs, args, predict, ncoefs, varimp) {
  new("ENMdetails", name = name, fun = fun, errors = errors, msgs = msgs, 
      args = args, predict = predict, ncoefs = ncoefs, varimp = varimp)
}
}
