% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{enm.ncoefs}
\alias{enm.ncoefs}
\alias{enm.ncoefs<-}
\alias{enm.ncoefs,ENMdetails-method}
\alias{enm.ncoefs<-,ENMdetails-method}
\title{enm.ncoefs generic for ENMdetails object}
\usage{
enm.ncoefs(x)

enm.ncoefs(x) <- value

\S4method{enm.ncoefs}{ENMdetails}(x)

\S4method{enm.ncoefs}{ENMdetails}(x) <- value
}
\arguments{
\item{x}{ENMdetails object}

\item{value}{input value}
}
\description{
enm.ncoefs generic for ENMdetails object
}
