% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{enm.predict}
\alias{enm.predict}
\alias{enm.predict<-}
\alias{enm.predict,ENMdetails-method}
\alias{enm.predict<-,ENMdetails-method}
\title{enm.predict generic for ENMdetails object}
\usage{
enm.predict(x)

enm.predict(x) <- value

\S4method{enm.predict}{ENMdetails}(x)

\S4method{enm.predict}{ENMdetails}(x) <- value
}
\arguments{
\item{x}{ENMdetails object}

\item{value}{input value}
}
\description{
enm.predict generic for ENMdetails object
}
