% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{enm.variable.importance}
\alias{enm.variable.importance}
\alias{enm.variable.importance<-}
\alias{enm.variable.importance,ENMdetails-method}
\alias{enm.variable.importance<-,ENMdetails-method}
\title{enm.variable.importance generic for ENMdetails object}
\usage{
enm.variable.importance(x)

enm.variable.importance(x) <- value

\S4method{enm.variable.importance}{ENMdetails}(x)

\S4method{enm.variable.importance}{ENMdetails}(x) <- value
}
\arguments{
\item{x}{ENMdetails object}

\item{value}{input value}
}
\description{
enm.variable.importance generic for ENMdetails object
}
