% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{enm.varimp}
\alias{enm.varimp}
\alias{enm.varimp<-}
\alias{enm.varimp,ENMdetails-method}
\alias{enm.varimp<-,ENMdetails-method}
\title{enm.varimp generic for ENMdetails object}
\usage{
enm.varimp(x)

enm.varimp(x) <- value

\S4method{enm.varimp}{ENMdetails}(x)

\S4method{enm.varimp}{ENMdetails}(x) <- value
}
\arguments{
\item{x}{ENMdetails object}

\item{value}{input value}
}
\description{
enm.varimp generic for ENMdetails object
}
