% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{enmeval.plot.grps}
\alias{enmeval.plot.grps}
\title{Partition group plots}
\usage{
enmeval.plot.grps(
  e = NULL,
  envs,
  pts = NULL,
  pts.grp = NULL,
  pts.type = "occs"
)
}
\arguments{
\item{e}{ENMevaluation object}

\item{envs}{Raster of an environmental predictor variable used to build the models in "e"}

\item{pts}{Matrix or data frame of coordinates for occurrence or background data}

\item{pts.grp}{Numeric vector of partition groups corresponding to data in "pts"}

\item{pts.type}{Character specifying which to plot: occurrences ("occs") or background ("bg"), with default "occs"}
}
\description{
Plot occurrence partition groups over an environmental predictor raster.
}
\details{
This function serves as a quick way to visualize occurrence or background partitions over the extent of an environmental predictor raster.
It can be run with an existing ENMevaluate object, or alternatively with occurrence or background coordinates and the corresponding partitions.
}
