% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enmeval_results.R
\docType{data}
\name{enmeval_results}
\alias{enmeval_results}
\title{Example ENMevaluation object.}
\format{
An ENMevaluation object
}
\usage{
enmeval_results
}
\description{
An example ENMevaluation object produced after running ENMevaluate() with 
feature classes L, LQ, LQH, H and regularization multipliers 1 through 5.
The SpatRaster data in the predictions slot is wrapped with `terra::wrap` to 
preserve the connections to the raster data. To use it, do the following:
`e@predictions <- terra::unwrap(e@predictions)`.
}
\keyword{datasets}
