\name{ENMevaluation-class }
\Rdversion{1.1}
\docType{class}
\alias{ENMevaluation-class}
\alias{ENMevaluation}
\title{Class \code{"ENMevaluation"}}
\description{ Objects of this class are generated by a call of \code{\link{ENMevaluate}}. }
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("ENMevaluation", ...)}.
}
\section{Slots}{
  \describe{
    \item{\code{results}:}{Object of class \code{"data.frame"} ~~ }
    \item{\code{predictions}:}{Object of class \code{"RasterStack"} ~~ }
    \item{\code{partition.method}:}{Object of class \code{"character"} ~~ }
    \item{\code{occ.pts}:}{Object of class \code{"data.frame"} ~~ }
    \item{\code{occ.grp}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{bg.pts}:}{Object of class \code{"data.frame"} ~~ }
    \item{\code{bg.grp}:}{Object of class \code{"numeric"} ~~ }
    \item{\code{overlap}:}{Object of class \code{"matrix"} ~~ }
  }
}
\author{ Jamie M. Kass <jkass@gc.cuny.edu> and Robert Muscarella <bob.muscarella@gmail.com> }
\examples{
showClass("ENMevaluation")
}
\keyword{classes}
