% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ENMevaluation_convert}
\alias{ENMevaluation_convert}
\title{Convert old ENMevaluation objects to new ones}
\usage{
ENMevaluation_convert(e, envs)
}
\arguments{
\item{e}{ENMevaluation object: the old object to convert}

\item{envs}{RasterStack: the original predictor variables used to generate
the old ENMevaluation object (these are used to make the new occs and bg slots
which contain the predictor variable values)}
}
\description{
Converts ENMevaluation objects made with version <=0.3.1 to
new ones made with version >=2.0.0.
}
\note{
If bin.output was set to TRUE, \code{`e@results`} will be equivalent to 
the new results.partitions slot. Some slots are unable to be filled in because
previous versions of ENMeval did not record them in ENMevaluation objects:
varimp, partition.settings, other.settings, doClamp (set to TRUE
arbitrarily to avoid errors, but may actually have been FALSE), clamp.directions,
taxon.name, and rmm.
}
