% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENMnullSims.R
\name{ENMnullSims}
\alias{ENMnullSims}
\title{Generate null ecological niche models (ENMs) and compare evaluation statistics with real ENMs}
\usage{
ENMnullSims(
  e,
  mod.settings,
  no.iter,
  envs = NULL,
  user.enm = NULL,
  user.envs.partition = NULL,
  userStats.signs = NULL,
  removeMxTemp = TRUE
)
}
\arguments{
\item{e}{ENMevaluation object}

\item{mod.settings}{named list of one set of model settings to build null models with}

\item{no.iter}{numeric of number of null model iterations}

\item{envs}{Raster* object of environmental variables (must be in same geographic projection as occurrence data);
necessary when evaluating using spatial cross-validation (see details)}

\item{user.enm}{ENMdetails object specified by the user}

\item{user.envs.partition}{numeric vector designating user-defined partition groups (folds) for \code{envs}; 
necessary when evaluating using spatial cross-validation (see details)}

\item{userStats.signs}{named list of user-defined evaluation statistics attributed with
either 1 or -1 to designate whether the expected difference between real and null models is 
positive or negative; this is used to calculate the p-value of the z-score}

\item{removeMxTemp}{boolean (TRUE or FALSE) which if TRUE delete all temporary data generated
when using maxent.jar for modeling}
}
\description{
\code{ENMnullSims()} builds null ecological niche models iteratively for a single set of 
user-specified model settings based on an input ENMevaluation object, from which all other analysis 
settings are copied. \code{ENMnullSims()} returns an \code{ENMnull} object with slots containing evaluation
summary statistics for the null models and their cross-validation results, as well as differences
in results between the real and null models. This comparison table includes z-scores of these differences
and their associated p-values (under a normal distribution).
}
