% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{eval.algorithm}
\alias{eval.algorithm}
\alias{eval.algorithm,ENMevaluation-method}
\title{eval.algorithm generic for ENMevaluation object}
\usage{
eval.algorithm(x)

\S4method{eval.algorithm}{ENMevaluation}(x)
}
\arguments{
\item{x}{ENMevaluation object}
}
\description{
eval.algorithm generic for ENMevaluation object
}
