% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{eval.bg.grp}
\alias{eval.bg.grp}
\alias{eval.bg.grp,ENMevaluation-method}
\title{eval.bg.grp generic for ENMevaluation object}
\usage{
eval.bg.grp(x)

\S4method{eval.bg.grp}{ENMevaluation}(x)
}
\arguments{
\item{x}{ENMevaluation object}
}
\description{
eval.bg.grp generic for ENMevaluation object
}
