% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{eval.bg}
\alias{eval.bg}
\alias{eval.bg,ENMevaluation-method}
\title{eval.bg generic for ENMevaluation object}
\usage{
eval.bg(x)

\S4method{eval.bg}{ENMevaluation}(x)
}
\arguments{
\item{x}{ENMevaluation object}
}
\description{
eval.bg generic for ENMevaluation object
}
