% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{eval.clamp.directions}
\alias{eval.clamp.directions}
\alias{eval.clamp.directions,ENMevaluation-method}
\title{eval.clamp.directions generic for ENMevaluation object}
\usage{
eval.clamp.directions(x)

\S4method{eval.clamp.directions}{ENMevaluation}(x)
}
\arguments{
\item{x}{ENMevaluation object}
}
\description{
eval.clamp.directions generic for ENMevaluation object
}
