% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{eval.doClamp}
\alias{eval.doClamp}
\alias{eval.doClamp,ENMevaluation-method}
\title{eval.doClamp generic for ENMevaluation object}
\usage{
eval.doClamp(x)

\S4method{eval.doClamp}{ENMevaluation}(x)
}
\arguments{
\item{x}{ENMevaluation object}
}
\description{
eval.doClamp generic for ENMevaluation object
}
