% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{eval.models}
\alias{eval.models}
\alias{eval.models,ENMevaluation-method}
\alias{eval.varimp,ENMevaluation-method}
\title{eval.models generic for ENMevaluation object}
\usage{
eval.models(x)

\S4method{eval.models}{ENMevaluation}(x)

\S4method{eval.varimp}{ENMevaluation}(x)
}
\arguments{
\item{x}{ENMevaluation object}
}
\description{
eval.models generic for ENMevaluation object
}
