% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{eval.occs.grp}
\alias{eval.occs.grp}
\alias{eval.occs.grp,ENMevaluation-method}
\title{eval.occs.grp generic for ENMevaluation object}
\usage{
eval.occs.grp(x)

\S4method{eval.occs.grp}{ENMevaluation}(x)
}
\arguments{
\item{x}{ENMevaluation object}
}
\description{
eval.occs.grp generic for ENMevaluation object
}
