% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{eval.occs}
\alias{eval.occs}
\alias{eval.occs,ENMevaluation-method}
\title{eval.occs generic for ENMevaluation object}
\usage{
eval.occs(x)

\S4method{eval.occs}{ENMevaluation}(x)
}
\arguments{
\item{x}{ENMevaluation object}
}
\description{
eval.occs generic for ENMevaluation object
}
