% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{eval.occs.testing}
\alias{eval.occs.testing}
\alias{eval.occs.testing,ENMevaluation-method}
\title{eval.occs.testing generic for ENMevaluation object}
\usage{
eval.occs.testing(x)

\S4method{eval.occs.testing}{ENMevaluation}(x)
}
\arguments{
\item{x}{ENMevaluation object}
}
\description{
eval.occs.testing generic for ENMevaluation object
}
