% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{eval.other.settings}
\alias{eval.other.settings}
\alias{eval.other.settings,ENMevaluation-method}
\title{eval.other.settings generic for ENMevaluation object}
\usage{
eval.other.settings(x)

\S4method{eval.other.settings}{ENMevaluation}(x)
}
\arguments{
\item{x}{ENMevaluation object}
}
\description{
eval.other.settings generic for ENMevaluation object
}
