% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{eval.overlap}
\alias{eval.overlap}
\alias{eval.overlap,ENMevaluation-method}
\title{eval.overlap generic for ENMevaluation object}
\usage{
eval.overlap(x)

\S4method{eval.overlap}{ENMevaluation}(x)
}
\arguments{
\item{x}{ENMevaluation object}
}
\description{
eval.overlap generic for ENMevaluation object
}
