% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{eval.partition.method}
\alias{eval.partition.method}
\alias{eval.partition.method,ENMevaluation-method}
\title{eval.partition.method generic for ENMevaluation object}
\usage{
eval.partition.method(x)

\S4method{eval.partition.method}{ENMevaluation}(x)
}
\arguments{
\item{x}{ENMevaluation object}
}
\description{
eval.partition.method generic for ENMevaluation object
}
