% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{eval.partition.settings}
\alias{eval.partition.settings}
\alias{eval.partition.settings,ENMevaluation-method}
\title{eval.partition.settings generic for ENMevaluation object}
\usage{
eval.partition.settings(x)

\S4method{eval.partition.settings}{ENMevaluation}(x)
}
\arguments{
\item{x}{ENMevaluation object}
}
\description{
eval.partition.settings generic for ENMevaluation object
}
