% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{eval.predictions}
\alias{eval.predictions}
\alias{eval.predictions,ENMevaluation-method}
\title{eval.predictions generic for ENMevaluation object}
\usage{
eval.predictions(x)

\S4method{eval.predictions}{ENMevaluation}(x)
}
\arguments{
\item{x}{ENMevaluation object}
}
\description{
eval.predictions generic for ENMevaluation object
}
