% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{eval.results.partitions}
\alias{eval.results.partitions}
\alias{eval.results.partitions,ENMevaluation-method}
\title{eval.results.partitions generic for ENMevaluation object}
\usage{
eval.results.partitions(x)

\S4method{eval.results.partitions}{ENMevaluation}(x)
}
\arguments{
\item{x}{ENMevaluation object}
}
\description{
eval.results.partitions generic for ENMevaluation object
}
