% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{eval.results}
\alias{eval.results}
\alias{eval.results,ENMevaluation-method}
\title{eval.results generic for ENMevaluation object}
\usage{
eval.results(x)

\S4method{eval.results}{ENMevaluation}(x)
}
\arguments{
\item{x}{ENMevaluation object}
}
\description{
eval.results generic for ENMevaluation object
}
