% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{eval.rmm}
\alias{eval.rmm}
\alias{eval.rmm,ENMevaluation-method}
\title{eval.rmm generic for ENMevaluation object}
\usage{
eval.rmm(x)

\S4method{eval.rmm}{ENMevaluation}(x)
}
\arguments{
\item{x}{ENMevaluation object}
}
\description{
eval.rmm generic for ENMevaluation object
}
