% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{eval.taxon.name}
\alias{eval.taxon.name}
\alias{eval.taxon.name,ENMevaluation-method}
\title{eval.taxon.name generic for ENMevaluation object}
\usage{
eval.taxon.name(x)

\S4method{eval.taxon.name}{ENMevaluation}(x)
}
\arguments{
\item{x}{ENMevaluation object}
}
\description{
eval.taxon.name generic for ENMevaluation object
}
