% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{eval.variable.importance}
\alias{eval.variable.importance}
\title{eval.variable.importance generic for ENMevaluation object}
\usage{
eval.variable.importance(x)
}
\arguments{
\item{x}{ENMevaluation object}
}
\description{
eval.variable.importance generic for ENMevaluation object
}
