% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{eval.varimp}
\alias{eval.varimp}
\title{eval.varimp (variable importance) generic for ENMevaluation object}
\usage{
eval.varimp(x)
}
\arguments{
\item{x}{ENMevaluation object}
}
\description{
eval.varimp (variable importance) generic for ENMevaluation object
}
