% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{evalplot.all.density}
\alias{evalplot.all.density}
\title{Density plots for All Variables with Shared Y-Axis}
\usage{
evalplot.all.density(e, envs, var, bw.envs = 10)
}
\arguments{
\item{e}{An ENMevaluation object.}

\item{envs}{Raster data (SpatRaster) of environmental variables for model projection.}

\item{var}{A character string specifying the variable name for the response curve.}

\item{bw.envs}{The smoothing bandwidth to be used in the environmental variables}
}
\value{
A combined patchwork plot of all response curves with a shared y-axis label.
}
\description{
A wrapper function to plot response curves for all contributing variables and combine
them using patchwork. The plots share a common y-axis label.
}
\examples{
\dontrun{
occs <- read.csv(file.path(system.file(package="predicts"), "/ex/bradypus.csv"))[,2:3]
envs <- rast(list.files(path=paste(system.file(package="predicts"), "/ex", sep=""), 
                        pattern="tif$", full.names=TRUE))
# No biome
envs <- envs[[!(names(envs) \%in\% "biome")]]
occs.z <- cbind(occs, terra::extract(envs, occs, ID = FALSE))
bg <- as.data.frame(predicts::backgroundSample(envs, n = 10000))
names(bg) <- names(occs)
bg.z <- cbind(bg, terra::extract(envs, bg, ID = FALSE))
os <- list(abs.auc.diff = FALSE, pred.type = "cloglog", validation.bg = "partition")
ps <- list(orientation = "lat_lat")
e.maxnet <- ENMevaluate(occs, envs, bg, 
                       tune.args = list(fc = "LQ", rm = 1), 
                        partitions = "block", other.settings = os, partition.settings = ps,
                        algorithm = "maxnet", overlap = TRUE)
# Transfer envs
tr_envs <- envs * 1.5
# Plot
evalplot.all.density(e = e.maxent, envs = tr_envs)
}
}
\author{
Gonzalo E. Pinilla-Buitrago
}
