% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{evalplot.grps.envSim}
\alias{evalplot.grps.envSim}
\title{MESS plots for partition groups}
\usage{
evalplot.grps.envSim(
  envs,
  occs = NULL,
  bg = NULL,
  occs.grp = NULL,
  bg.grp = NULL,
  e = NULL,
  envs.var = NULL,
  sim.type = "mess",
  pts.type = "occs",
  plot.type = "histogram",
  plot.palette = NULL,
  palette.dir = 1,
  hist.bins = 30,
  return.output = FALSE
)
}
\arguments{
\item{envs}{RasterStack: environmental predictor variables used to build the models in "e"; categorical variables should be 
removed before input, as they cannot be used to calculate MESS}

\item{e}{ENMevaluation object}

\item{envs.var}{Character: the name of a predictor variable to plot similarities for; if left NULL, calculations are done
with respect to all variables}

\item{sim.type}{Character: either "mess" for Multivariate Environmental Similarity Surface, "most_diff" for most different variable,
or "most_sim" for most similar variable; uses similarity function from package rmaxent}

\item{pts.type}{Character: the reference to calculate MESS based on: occurrences ("occs") or background ("bg"), with default "occs"}

\item{plot.type}{Character specifying which to plot: MESS histograms ("histogram") or MESS rasters ("raster"); default is histogram}

\item{plot.palette}{Character: the RColorBrewer palette name to use for plotting; if left NULL, the defaults are
"Set1" for discrete variables and reverse "RdYlBu" for continuous variables}

\item{palette.dir}{Numeric: direction for palette plotting; either 1 or -1}

\item{hist.bins}{Numeric: number of histogram bins for histogram plots; default is 30}
}
\value{
A ggplot of MESS calculations for data partitions.
}
\description{
Plot environmental similarity of predictor variable extent with respect to occurrence partitions
}
\details{
There are two variations for this plot. If "histogram", histograms are plotted showing the MESS estimates for each partition group. 
If "raster", rasters are plotted showing the geographical MESS estimates for each partition group. 
With sim.type option "mess", the similarity between environmental values associated with the 
test occurrences (per partition group) and those associated with the entire study extent (specified by the extent 
of the input RasterStack "envs") are calculated, and the minimum similarity per grid is returned. 
Higher negative values indicate greater environmental difference between the test occurrences
and the study extent, and higher positive values indicate greater similarity. This function uses the `rmaxent::similarity()` function 
to calculate the similarities. See the below reference for details on MESS.
}
\references{
Elith J., M. Kearney M., and S. Phillips, 2010. The art of modelling range-shifting species. Methods in Ecology and Evolution 1:330-342.
}
