% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{evalplot.nulls}
\alias{evalplot.nulls}
\title{ENMnulls statistics plot}
\usage{
evalplot.nulls(
  e.null,
  stats,
  plot.type,
  facet.labels = NULL,
  metric.levels = NULL,
  return.tbl = FALSE
)
}
\arguments{
\item{e.null}{ENMnull object}

\item{stats}{Character vector: metrics from results table to be plotted; examples are
"auc.val" or "or.10p"; if more than one statistic is specified, the histogram plot will be faceted}

\item{plot.type}{Character: either "violin" or "histogram"}

\item{facet.labels}{Character vector: custom names for the metric facets}

\item{metric.levels}{Character vector: custom factor levels for metrics; this controls
the order that metric statistics are plotted}
}
\value{
A ggplot of null model statistics.
}
\description{
Plot histogram of evaluation statistics for null ENM simulations
}
\details{
There are two variations for this plot, but both show null quantiles (0.01, 0.05, 0.5, 0.95, 0.99). 
For violin plots, the null distribution is displayed as a vertical shape (i.e., the violin) with horizontal lines showing 
the quantiles and the empirical value is plotted as a red point along the vertical axis. 
For histogram plots, the null distribution is displayed as a histogram with vertical lines showing the quantiles 
and the empirical value as a vertical red line on the distribution.
}
