% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{loadENMevaluation}
\alias{loadENMevaluation}
\title{Load ENMevaluation object}
\usage{
loadENMevaluation(filename)
}
\description{
Load an ENMevaluation object as an .rds file. This is necessary
to use instead of readRDS() because wrapped terra SpatRasters require 
unwrap() after loading for the raster data. This convenience function does 
that for you.
}
