% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lookup.enm}
\alias{lookup.enm}
\title{Look up ENMdetails abject}
\usage{
lookup.enm(algorithm)
}
\arguments{
\item{algorithm}{character: algorithm name (must be implemented as ENMdetails object)}
}
\description{
Internal function to look up ENMdetails objects.
}
