% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{lookup.var.names}
\alias{lookup.var.names}
\title{Look up predictor variable names from a fitted model}
\usage{
lookup.var.names(mod, algorithm)
}
\arguments{
\item{mod}{A fitted model object.}

\item{algorithm}{character: algorithm name, as returned by [lookup.algorithm()].}
}
\description{
Internal function to get the predictor variable names used to fit
a model, given its algorithm.
}
