% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{maxnet.predictRaster}
\alias{maxnet.predictRaster}
\title{Predict raster for maxnet}
\usage{
maxnet.predictRaster(m, envs, type = "cloglog", doClamp = TRUE, other.settings)
}
\description{
As maxnet does not have native functionality for making 
prediction rasters, this function does it. It is a wrapper for the internal
enm.maxnet@predict function, which is not really intended for use outside
the package.
}
