% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mess.vec}
\alias{mess.vec}
\title{Compute multivariate environmental similarity surfaces (MESS)}
\usage{
mess.vec(p, v)
}
\arguments{
\item{p}{Raster object}

\item{v}{Matrix with reference values}
}
\value{
A RasterBrick with layers corresponding to the input layers and an additional layer with the mess values (if full=TRUE and nlayers(x) > 1) or a RasterLayer with the MESS values (if full=FALSE).
}
\description{
Compute multivariate environmental similarity surfaces (MESS) (i.e., Elith \emph{et al.} 2010).
}
\details{
Repurposed from dismo::mess(), based on .messi3()
}
\references{
Elith J., M. Kearney M., and Phillips, S. (2010) The art of modelling range-shifting species. \emph{Methods in Ecology and Evolution}, \bold{1}: 330-342.
}
\author{
Based on \pkg{dismo}::\code{mess}
Jean-Pierre Rossi <jean-pierre.rossi@supagro.inra.fr>, Robert Hijmans, Paulo van Breugel
}
