% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{multiRasMatchNAs}
\alias{multiRasMatchNAs}
\title{convert all raster grid cells to NA that are NA for any one raster in
a multiraster SpatRaster object}
\usage{
multiRasMatchNAs(envs, quiet = TRUE)
}
\arguments{
\item{envs}{SpatRaster object}
}
\description{
match all NA cells in SpatRaster object
}
