% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{null.algorithm}
\alias{null.algorithm}
\alias{null.algorithm,ENMnull-method}
\title{null.algorithm generic for ENMnull object}
\usage{
null.algorithm(x)

\S4method{null.algorithm}{ENMnull}(x)
}
\arguments{
\item{x}{ENMnull object}
}
\description{
null.algorithm generic for ENMnull object
}
