% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{null.doClamp}
\alias{null.doClamp}
\alias{null.doClamp,ENMnull-method}
\title{null.doClamp generic for ENMnull object}
\usage{
null.doClamp(x)

\S4method{null.doClamp}{ENMnull}(x)
}
\arguments{
\item{x}{ENMnull object}
}
\description{
null.doClamp generic for ENMnull object
}
