% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{null.emp.results}
\alias{null.emp.results}
\alias{null.emp.results,ENMnull-method}
\title{null.emp.results generic for ENMnull object}
\usage{
null.emp.results(x)

\S4method{null.emp.results}{ENMnull}(x)
}
\arguments{
\item{x}{ENMnull object}
}
\description{
null.emp.results generic for ENMnull object
}
