% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{null.mod.settings}
\alias{null.mod.settings}
\alias{null.mod.settings,ENMnull-method}
\title{null.mod.settings generic for ENMnull object}
\usage{
null.mod.settings(x)

\S4method{null.mod.settings}{ENMnull}(x)
}
\arguments{
\item{x}{ENMnull object}
}
\description{
null.mod.settings generic for ENMnull object
}
