% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{null.no.iter}
\alias{null.no.iter}
\alias{null.no.iter,ENMnull-method}
\title{null.no.iter generic for ENMnull object}
\usage{
null.no.iter(x)

\S4method{null.no.iter}{ENMnull}(x)
}
\arguments{
\item{x}{ENMnull object}
}
\description{
null.no.iter generic for ENMnull object
}
