% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{null.other.settings}
\alias{null.other.settings}
\alias{null.other.settings,ENMnull-method}
\title{null.other.settings generic for ENMnull object}
\usage{
null.other.settings(x)

\S4method{null.other.settings}{ENMnull}(x)
}
\arguments{
\item{x}{ENMnull object}
}
\description{
null.other.settings generic for ENMnull object
}
