% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{null.partition.method}
\alias{null.partition.method}
\alias{null.partition.method,ENMnull-method}
\title{null.partition.method generic for ENMnull object}
\usage{
null.partition.method(x)

\S4method{null.partition.method}{ENMnull}(x)
}
\arguments{
\item{x}{ENMnull object}
}
\description{
null.partition.method generic for ENMnull object
}
