% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{null.results.partitions}
\alias{null.results.partitions}
\alias{null.results.partitions,ENMnull-method}
\title{null.results.partitions generic for ENMnull object}
\usage{
null.results.partitions(x)

\S4method{null.results.partitions}{ENMnull}(x)
}
\arguments{
\item{x}{ENMnull object}
}
\description{
null.results.partitions generic for ENMnull object
}
