% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{null.results}
\alias{null.results}
\alias{null.results,ENMnull-method}
\title{null.results generic for ENMnull object}
\usage{
null.results(x)

\S4method{null.results}{ENMnull}(x)
}
\arguments{
\item{x}{ENMnull object}
}
\description{
null.results generic for ENMnull object
}
