% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.dataPrep}
\alias{plot.dataPrep}
\title{Internal plotting function}
\usage{
\method{plot}{dataPrep}(data, envs, occs.z, bg.z, occs.grp, bg.grp, ref.data, occs.testing.z)
}
\description{
This function preps data for plotting.
}
\examples{
\dontrun{
plot.sim.dataPrep()
}
}
\keyword{internal}
