% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.envSim}
\alias{plot.envSim}
\title{Similarity histogram plots for partition groups}
\usage{
\method{plot}{envSim}(
  data = NULL,
  ref.data = "occs",
  lat.name = NULL,
  lon.name = NULL,
  part.name = NULL,
  type.name = NULL,
  envs.vars = NULL,
  occs.testing.z = NULL,
  hist.bins = 30,
  return.tbl = FALSE
)
}
\arguments{
\item{ref.data}{character: the reference to calculate MESS based on occurrences ("occs") or background ("bg"), with default "occs"
these must be specified as this function was intended for use with continuous data only; these must be specified when inputting tabular data instead of an ENMevaluation object}

\item{envs.vars}{character vector: names of a predictor variable to plot similarities for; if left NULL, calculations are done
with respect to all variables (optional)}

\item{occs.testing.z}{data frame: longitude, latitude, and environmental predictor variable values for fully withheld testing records, 
in that order; this is for use only with the "testing" partition option when an ENMevaluation object is not input (optional)}

\item{hist.bins}{numeric: number of histogram bins for histogram plots; default is 30}

\item{return.tbl}{boolean: if TRUE, return the data frames of similarity values used to make the ggplot instead of the plot itself}

\item{e}{ENMevaluation object}

\item{occs.z}{data frame: longitude, latitude, and environmental predictor variable values for occurrence records, in that order (optional);
the first two columns must be named "longitude" and "latitude"}

\item{occs.grp}{numeric vector: partition groups for occurrence records (optional)}

\item{bg.z}{data frame: longitude, latitude, and environmental predictor variable values for background records, in that order (optional);
the first two columns must be named "longitude" and "latitude"}

\item{bg.grp}{numeric vector: partition groups for background records (optional)}

\item{quiet}{boolean: if TRUE, silence all function messages (but not errors)}
}
\value{
A ggplot of environmental similarities between the occurrence or background data 
for each partition and the rest of the data (all other occurrences and background data).
}
\description{
Plots environmental similarity of reference partitions (occurrences or 
background) to remaining data (occurrence and background for all other partitions). This 
function does not use raster data, and thus only calculates similarity values for data used 
in model training. Further, this function does not calculate similarity for categorical 
variables.
}
\details{
When fully withheld testing groups are used, make sure to input either an 
ENMevaluation object or the argument occs.testing.z. In the resulting plot, partition 1 
refers to the training data, while partition 2 refers to the fully withheld testing group.

Histograms are plotted showing the environmental similarity estimates for each 
partition group in relation to the others. The similarity between environmental values associated with the 
validation occurrence or background records per partition group and those associated with 
the remaining data (training occurrences and background) are calculated with the MESS algorithm, and the minimum similarity 
per grid cell is returned. Higher negative values indicate a greater environmental difference between the validation occurrences and the study extent, and higher 
positive values indicate greater similarity. This function uses the `mess()` function 
from the package `predicts`. Please see the below reference for details on MESS.
}
\examples{
\dontrun{
# first, let's tune some models
occs <- read.csv(file.path(system.file(package="predicts"), 
"/ex/bradypus.csv"))[,2:3]
envs <- rast(list.files(path=paste(system.file(package="predicts"), 
"/ex", sep=""), pattern="tif$", full.names=TRUE))
bg <- as.data.frame(predicts::backgroundSample(envs, n = 10000))
names(bg) <- names(occs)
 
ps <- list(orientation = "lat_lat")

e <- ENMevaluate(occs, envs, bg, 
               tune.args = list(fc = c("L","LQ","LQH"), rm = 1:5), 
               partitions = "block", partition.settings = ps, 
               algorithm = "maxnet", categoricals = "biome", 
               parallel = TRUE)

# now, plot the environmental similarity of each partition to the others               
evalplot.envSim.hist(e)
}

}
\references{
Baumgartner J, Wilson P (2021). _rmaxent: Tools for working with Maxent in R_. R package version 0.8.5.9000, <URL: https://github.com/johnbaums/rmaxent>.
Elith, J., Kearney, M., and Phillips, S. (2010) The art of modelling range-shifting species. \emph{Methods in Ecology and Evolution}, \bold{1}: 330-342. \doi{doi:10.1111/j.2041-210X.2010.00036.x}
}
