% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.eval}
\alias{plot.eval}
\title{ENMevaluation statistics plot}
\usage{
\method{plot}{eval}(e, stats, x, col, dodge = NULL, error.bars = TRUE)
}
\arguments{
\item{e}{ENMevaluation object}

\item{stats}{vector of names of statistics from results table to be plotted; if more than
one statistic is specified, the plot will be faceted}

\item{x}{character of variable to be plotted on x-axis}

\item{col}{character of variable used to assign symbology colors}
}
\value{
A plot of evaluation statistics, with x representing a tuning parameter range, y
representing the average of a statistic over all partitions, and colors representing another
tuning parameter's values. Error bars represent the standard deviation of a statistic around the 
mean. Currently, this function only can handle two tuning parameters at a time.
}
\description{
Plot evaluation statistics over tuning parameter ranges to visualize differences in performance
}
