% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.grps}
\alias{plot.grps}
\title{Partition group plots}
\usage{
plot.grps(e = NULL, pts = NULL, pts.grp = NULL, envs, pts.type = "occs")
}
\arguments{
\item{e}{ENMevaluation object}

\item{envs}{Raster of an environmental predictor variable used to build the models in "e"}

\item{pts.type}{character specifying which to plot: occurrences ("occs") or background ("bg"), with default "occs"}
}
\description{
Plot partition groups over an environmental predictor raster
}
