% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.nullENMs}
\alias{plot.nullENMs}
\title{ENMnullSims statistics plot}
\usage{
\method{plot}{nullENMs}(e.null, stats, plot.type)
}
\arguments{
\item{e.null}{ENMnull object}

\item{stats}{vector of names of statistics from results table to be plotted; if more than
one statistic is specified, the histogram plot will be faceted}

\item{plot.type}{either "violin" or "histogram"}
}
\value{
A plot of evaluation statistics for null ENM simulations and display positions of quantiles and real value.
}
\description{
Plot histogram of evaluation statistics for null ENM simulations
}
