% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.maxnet.R
\name{response.maxnet}
\alias{response.maxnet}
\title{Generate values needed to plot a response curve from a maxnet model object}
\usage{
response.maxnet(mod, v, type)
}
\arguments{
\item{mod}{maxnet model object}

\item{v}{character of variable name}

\item{type}{character of prediction type: one of "link", "exponential", "cloglog", or "logistic"}
}
\value{
Matrix with first column as variable values (x-axis), and second column as predicted values (y-axis).
}
\description{
Generate values needed to plot a response curve from a maxnet model object
}
