% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{saveENMevaluation}
\alias{saveENMevaluation}
\title{Save ENMevaluation object}
\usage{
saveENMevaluation(e, filename)
}
\arguments{
\item{e}{ENMevaluation object}

\item{filename}{character: path to the file to create with .rds extension}
}
\description{
Save an ENMevaluation object as an .rds file. This is necessary
to use instead of \code{saveRDS()} because terra SpatRasters require \code{wrap()} before
saving to preserve the connections to the raster data. This convenience 
function does that for you.
}
