% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{user.eval}
\alias{user.eval}
\title{User-specified evaluation function for model validation}
\usage{
my.user.eval <- myFunc(vars)
ENMevaluate(..., user.eval = my.user.eval)
}
\arguments{
\item{enm}{ENMdetails object}

\item{occs.train.z}{data frame: predictor variable values for training occurrences}

\item{occs.val.z}{data frame: predictor variable values for validation occurrences}

\item{bg.train.z}{data frame: predictor variable values for training background}

\item{bg.val.z}{data frame: predictor variable values for validation background}

\item{mod.k}{Model object for current partition (k)}

\item{nk}{numeric: number of folds (i.e., partitions)}

\item{other.settings}{named list: other settings specified in ENMevaluate()}

\item{partitions}{character: name of the partition method (e.g., "block")}

\item{occs.train.pred}{numeric: predictions made by mod.k for training occurrences}

\item{occs.val.pred}{numeric: predictions made by mod.k for validation occurrences}

\item{bg.train.pred}{numeric: predictions made by mod.k for training background}

\item{bg.val.pred}{numeric: predictions made by mod.k for validation background}
}
\description{
This is a custom function for specifying performance metrics not included in ENMeval.
The function must be first defined and then input as the argument "user.eval" into ENMevaluate(). 
This function has a single argument called "vars", which is a list that includes different data 
that can be used to calculate the metric. Below is the list of data included in vars, which can
be accessed with $, as in vars$occs.train.z. See the vignette for a worked example.
}
