% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{calc.10p.trainThresh}
\alias{calc.10p.trainThresh}
\title{Calculate 10 percentile threshold}
\usage{
calc.10p.trainThresh(pred.train)
}
\arguments{
\item{pred.train}{numeric vector: training occurrence predictions}
}
\description{
Function to calculate the 10 percentile threshold from training predictions
}
