% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{enm.name}
\alias{enm.name}
\alias{enm.name<-}
\alias{enm.name,ENMdetails-method}
\alias{enm.name<-,ENMdetails-method}
\title{eval.name generic for ENMdetails object}
\usage{
enm.name(x)

enm.name(x) <- value

\S4method{enm.name}{ENMdetails}(x)

\S4method{enm.name}{ENMdetails}(x) <- value
}
\arguments{
\item{x}{ENMdetails object}

\item{value}{input value}
}
\description{
eval.name generic for ENMdetails object
}
