% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enm.rf.R
\docType{data}
\name{enm.rf}
\alias{enm.rf}
\title{ENMdetails rf}
\format{
An object of class \code{ENMdetails} of length 1.
}
\usage{
enm.rf
}
\description{
This is the ENMdetails implementation for random forest, the R version of
the Maxent algorithm. The configuration for running the model now includes addsamplestobackground = TRUE,
which explicitly adds presences to the background for model training, though as the current 
version of rf has this set to TRUE as default, behavior between ENMeval versions should not differ.
}
\keyword{datasets}
