% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{ENMdetails}
\alias{ENMdetails}
\title{ENMdetails class}
\usage{
ENMdetails(name, fun, msgs, args, predict, ncoefs, varimp)
}
\description{
An S4 class that details packages, functions, messages associated with a specific species distribution model (SDM) or ecological niche model (ENM). 
Objects of this class are generated by \code{ENMdetails()}. For examples, look in the package's R folder for scripts beginning with "enm" -- these are 
pre-made ENMdetails object specifications that work with ENMeval out of the box.
}
\section{Slots}{

\describe{
\item{\code{name}}{character of name of algorithm}

\item{\code{fun.train}}{function that runs the training algorithm (can be same as validation algorithm)}

\item{\code{fun.val}}{function that runs the validation algorithm (can be same as training algorithm)}

\item{\code{msgs}}{function that prints messages showing the package version number, etc., and those related to the input tuning parameters \code{tune.args}}

\item{\code{args.train}}{function specifying the parameters needed to run the training model function (can be same as validation algorithm)}

\item{\code{args.val}}{function specifying the parameters needed to run the validation model function (can be same as training algorithm)}

\item{\code{predict}}{function specifying how to calculate a model prediction for a Raster* or a data frame}

\item{\code{ncoefs}}{function specifying how to measure the number of non-zero model coefficients}

\item{\code{varimp}}{function specifying how to generate a data frame of variable importance from the model object (if possible)}
}}

\author{
Jamie M. Kass, \email{jamie.m.kass@gmail.com}
}
