% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{ENMdetails}
\alias{ENMdetails}
\title{ENMdetails class}
\usage{
ENMdetails(
  name,
  fun,
  pkgs,
  msgs,
  args,
  aic,
  eval.train,
  eval.test,
  pred,
  nparams
)
}
\description{
An S4 class that details packages, functions, messages associated with a specific species distribution model (SDM) or ecological niche model (ENM). 
Objects of this class are generated by \code{ENMdetails()}. For examples, look in the package's R folder for scripts beginning with "enm" -- these are 
pre-made ENMdetails object specifications that work with ENMeval out of the box.
}
\section{Slots}{

\describe{
\item{\code{name}}{character of name of model}

\item{\code{fun}}{function that runs the model}

\item{\code{pkgs}}{vector of package names needed to run the model function}

\item{\code{msgs}}{function that prints messages showing the package version number, etc., and those related to the input tuning parameters \code{tune.args}}

\item{\code{args}}{function specifying the parameters needed to run the model function}

\item{\code{aic}}{function specifying how AIC should be calculated (if at all)}

\item{\code{eval.train}}{function specifying how to calculate training evaluation statistics}

\item{\code{eval.test}}{function specifying how to calculate testing evaluation statistics}

\item{\code{pred}}{function specifying how to calculate a model prediction for a Raster* or a data frame}

\item{\code{nparams}}{function specifying how to measure the number of model coefficients}
}}

\author{
Jamie M. Kass, \email{jamie.m.kass@gmail.com}
}
