% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENMeval-package.R
\docType{package}
\name{ENMeval-package}
\alias{ENMeval}
\alias{ENMeval-package}
\title{Automated runs and evaluations of ecological niche models}
\description{
Runs ecological niche models over all combinations of user-defined settings (i.e., tuning), performs cross validation to evaluate models, and returns data tables to aid in selection of optimal model settings that balancegoodness-of-fit and model complexity. Also has functions to partition data spatially (or not) for cross validation, to plot the tuning results, and to calculate niche overlap between model predictions, among others. Originally built for Maxent models (Phillips et al. 2006, Phillips et al. 2017), but current version allows possible extensions for any modeling algorithm.
}
\details{

}
\keyword{ENM}
\keyword{NULL}
\keyword{SDM}
\keyword{niche}
