% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{data}
\name{enmeval_results}
\alias{enmeval_results}
\title{An object of class `ENMevaluation`.}
\format{An object of class `ENMevaluation`.}
\source{
Simulated data from `ENMevaluate`.
}
\usage{
enmeval_results
}
\description{
An example results file based on a call of `ENMevaluate` (see example).
}
\details{
The dataset is based on the simulated dataset and call of \code{\link{ENMevaluate}} shown in the example section below.
}
\examples{
require(raster)
### Simulated data environmental covariates
set.seed(1)
r1 <- raster(matrix(nrow=50, ncol=50, data=runif(10000, 0, 25)))
r2 <- raster(matrix(nrow=50, ncol=50, data=rep(1:100, each=100), byrow=TRUE))
r3 <- raster(matrix(nrow=50, ncol=50, data=rep(1:100, each=100)))
r4 <- raster(matrix(nrow=50, ncol=50, data=c(rep(1,1000),rep(2,500)),byrow=TRUE))
values(r4) <- as.factor(values(r4))
env <- stack(r1,r2,r3,r4)

### Simulate occurrence localities
nocc <- 50
x <- (rpois(nocc, 2) + abs(rnorm(nocc)))/11
y <- runif(nocc, 0, .99)
occ <- cbind(x,y)
\dontrun{
enmeval_results <- ENMevaluate(occs, env, n.bg=500, 
                               partitions="block", 
                               categoricals="layer.4",
                               mod.name='maxnet', 
                               tune.args=list(fc = c("L","LQ","LQH","LQHP","LQHPT"), 
                                              rm = 1:4),
                               overlap=T, overlapStat="D")
}

data(enmeval_results)
enmeval_results

### See table of evaluation metrics
enmeval_results@results

### Plot prediction with lowest AICc
plot(enmeval_results@predictions[[which (enmeval_results@results$delta.AICc == 0) ]])
points(enmeval_results@occ.pts, pch=21, bg= enmeval_results@occ.grp)

### Niche overlap statistics between model predictions
enmeval_results@overlap
}
\keyword{datasets}
