% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENMevaluate.R
\name{ENMevaluate}
\alias{ENMevaluate}
\title{Tune ecological niche model (ENM) settings and calculate evaluation statistics}
\usage{
ENMevaluate(
  occs,
  envs = NULL,
  bg = NULL,
  tune.args = NULL,
  partitions = NULL,
  algorithm = NULL,
  partition.settings = NULL,
  other.settings = NULL,
  categoricals = NULL,
  user.enm = NULL,
  user.grp = NULL,
  occs.testing = NULL,
  taxon.name = NULL,
  n.bg = 10000,
  overlap = FALSE,
  overlapStat = c("D", "I"),
  user.val.grps = NULL,
  user.eval = NULL,
  rmm = NULL,
  parallel = FALSE,
  numCores = NULL,
  parallelType = "doSNOW",
  updateProgress = FALSE,
  quiet = FALSE,
  occ = NULL,
  env = NULL,
  bg.coords = NULL,
  RMvalues = NULL,
  fc = NULL,
  occ.grp = NULL,
  bg.grp = NULL,
  method = NULL,
  bin.output = NULL,
  rasterPreds = NULL,
  progbar = NULL
)
}
\arguments{
\item{occs}{matrix or data frame with three columns for taxon name, longitude, and latitude 
of occurrence localities, in that order; if specifying predictor variable values
assigned to presence/background localities (species with data "SWD" form), this table will also have 
one column for each predictor variable}

\item{envs}{Raster* object of environmental variables (must be in same geographic projection as occurrence data)}

\item{bg}{matrix or data frame with two columns for longitude and latitude of 
background (or pseudo-absence) localities, in that order; if specifying predictor variable values
assigned to presence/background localities (species with data "SWD" form), this table will also have 
one column for each predictor variable; if NULL, points will be randomly sampled across \code{envs} 
with the number specified by argument \code{n.bg}}

\item{tune.args}{named list of model settings to be tuned}

\item{partitions}{character of name of partitioning technique (see \code{?partitions})}

\item{algorithm}{character of the name of the chosen model}

\item{categoricals}{character vector of names of categorical environmental variables}

\item{user.enm}{ENMdetails object specified by the user; this model will be
used for the analysis, and is an alternative to specifying algorithm}

\item{user.grp}{named list with occs.grp = vector of partition group (fold) for each
occurrence locality, intended for user-defined partitions, and bg.grp = same vector for 
background (or pseudo-absence) localities}

\item{occs.testing}{matrix or data frame with two columns for longitude and latitude 
of occurrence localities, in that order, intended for a testing dataset;
when \code{partitions = "testing"}; these occurrences will be used only 
for evaluation, and not for model training, and thus no cross validation will be done}

\item{n.bg}{numeric (default: 10000) for number of random background (or pseudo-absence) points to
sample; necessary if \code{bg} is NULL}

\item{overlap}{boolean (TRUE or FALSE) which if TRUE, calculate niche overlap statistics}

\item{overlapStat}{character for one or two (vector) niche overlap statistics:
choices are "D" and "I" -- see ?calc.niche.overlap for more details}

\item{user.val.grps}{matrix or data frame of user-defined test record coordinates and predictor variable values; this is mainly used
internally by ENMnullSims() to force each null model to evaluate with real test data}

\item{user.eval}{function specifying custom validation evaluation (see vignette for example)}

\item{rmm}{RMM object to be written to by ENMevaluate; if not specified, a new RMM object is output}

\item{parallel}{boolean (TRUE or FALSE) which if TRUE, run with parallel processing}

\item{numCores}{numeric for number of cores to use for parallel processing}

\item{parallelType}{character (default: "doSNOW") specifying either "doParallel" or "doSNOW"}

\item{updateProgress}{boolean (TRUE or FALSE) which if TRUE, use shiny progress bar; only for use in shiny apps}

\item{other.args}{named list of any additional model arguments not specified for tuning}

\item{clamp}{boolean; if TRUE, model prediction extrapolations will be restricted to the upper and lower
bounds of the predictor variables -- this avoids extreme predictions for non-analog environments, but
if extrapolation is a study aim, this should be set to FALSE}
}
\value{

}
\description{
\code{ENMevaluate()} builds ecological niche models iteratively across a range of 
user-specified tuning settings. Users can choose to evaluate models with cross validation or a
full-withheld testing dataset. \code{ENMevaluate()} returns an \code{ENMevaluation} object with slots containing 
evaluation statistics for each combination of settings and for each cross validation fold therein, as
well as raster predictions for each model when raster data is input. The evaluation statistics in the 
results table should aid users in identifying model settings that balance fit and predictive ability.
}
\examples{

}
