% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{ENMevaluation-class}
\alias{ENMevaluation-class}
\alias{ENMevaluation}
\title{ENMevaluation class}
\description{
An S4 class that contains the ENMevaluate results.
}
\section{Slots}{

\describe{
\item{\code{algorithm}}{character of algorithm used}

\item{\code{tune.settings}}{data.frame of settings that were tuned}

\item{\code{partition.method}}{character of partition method used}

\item{\code{partition.settings}}{list of partition settings used (i.e., value of *k* or aggregation factor)}

\item{\code{other.settings}}{list of other modeling settings used (i.e., decisions about clamping, AUC diff calculation)}

\item{\code{results}}{data.frame of evaluation summary statistics}

\item{\code{results.partitions}}{data.frame of evaluation k-fold statistics}

\item{\code{models}}{list of model objects}

\item{\code{predictions}}{RasterStack of model predictions}

\item{\code{occs}}{data.frame of occurrence coordinates and predictor variable values used for model training}

\item{\code{occs.grp}}{vector of partition groups for occurrence points}

\item{\code{bg}}{data.frame of background coordinates and predictor variable values used for model training}

\item{\code{bg.grp}}{vector of partition groups for background points}

\item{\code{overlap}}{list of matrices of pairwise niche overlap statistics}
}}

\author{
Jamie M. Kass, \email{jamie.m.kass@gmail.com}, Bob Muscarella, \email{bob.muscarella@gmail.com}
}
