% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{ENMnull-class}
\alias{ENMnull-class}
\alias{ENMnull}
\title{ENMnull class}
\description{
An S4 class that contains the ENMnullSims results.
}
\section{Slots}{

\describe{
\item{\code{null.algorithm}}{character of algorithm used}

\item{\code{null.mod.settings}}{data.frame of model settings used}

\item{\code{null.partition.method}}{character of partition method used}

\item{\code{null.partition.settings}}{list of partition settings used (i.e., value of *k* or aggregation factor)}

\item{\code{null.other.settings}}{list of other modeling settings used (i.e., decisions about clamping, AUC diff calculation)}

\item{\code{no.iter}}{numeric of number of null model iterations}

\item{\code{null.results}}{data.frame of evaluation summary statistics for null models}

\item{\code{null.results.grp}}{data.frame of evaluation k-fold statistics for null models}

\item{\code{real.vs.null.results}}{data.frame of evaluation summary statistics for the real model, means for all null models, z-scores, and p-values}

\item{\code{real.occs}}{data.frame of occurrence coordinates and predictor variable values used for model training (real model)}

\item{\code{real.occs.grp}}{vector of partition groups for occurrence points}

\item{\code{real.bg}}{data.frame of background coordinates and predictor variable values used for model training}

\item{\code{real.bg.grp}}{vector of partition groups for background points}
}}

\author{
Jamie M. Kass, \email{jamie.m.kass@gmail.com}, Corentin Bohl, \email{corentinbohl@gmail.com}
}
