% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{evalplot.density}
\alias{evalplot.density}
\title{Plot Density Plots of variables}
\usage{
evalplot.density(data, envs, var, bw.envs = 10)
}
\arguments{
\item{data}{Data frame of training data (occurrences + background).}

\item{envs}{Raster data (SpatRaster) of environmental variables for model projection.}

\item{var}{A character string specifying the variable name for the response curve.}

\item{bw.envs}{The smoothing bandwidth to be used in the environmental variables}
}
\value{
A ggplot object of the response curve.
}
\description{
This function plots densities of a given environmental variable based on a maxent.jar
or maxnet model.
}
\examples{
\dontrun{
library(ENMeval)
occs <- read.csv(file.path(system.file(package="predicts"), "/ex/bradypus.csv"))[,2:3]
envs <- terra::rast(list.files(path=paste(system.file(package="predicts"), "/ex", sep=""), 
                        pattern="tif$", full.names=TRUE))
# No biome
envs <- envs[[!(names(envs) \%in\% "biome")]]
occs.z <- cbind(occs, terra::extract(envs, occs, ID = FALSE))
bg <- as.data.frame(predicts::backgroundSample(envs, n = 10000))
names(bg) <- names(occs)
bg.z <- cbind(bg, terra::extract(envs, bg, ID = FALSE))
os <- list(abs.auc.diff = FALSE, pred.type = "cloglog", validation.bg = "partition")
ps <- list(orientation = "lat_lat")
e.maxnet <- ENMevaluate(occs, envs, bg, 
                       tune.args = list(fc = "LQ", rm = 1), 
                        partitions = "block", other.settings = os, partition.settings = ps,
                        algorithm = "maxnet", overlap = TRUE)
# Transfer envs
tr_envs <- envs * 1.5
# Define data as combined training values with coordinates removed
data <- rbind(e@occs, e@bg)[,3:11]
# Plot
evalplot.density(data, envs = tr_envs, var = "bio1")
}
}
\author{
Gonzalo E. Pinilla-Buitrago
}
