% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{evalplot.envSim.hist}
\alias{evalplot.envSim.hist}
\title{Similarity histogram plots for partition groups}
\usage{
evalplot.envSim.hist(
  e = NULL,
  occs.z = NULL,
  bg.z = NULL,
  occs.grp = NULL,
  bg.grp = NULL,
  ref.data = "occs",
  sim.type = "mess",
  categoricals = NULL,
  envs.var = NULL,
  occs.testing.z = NULL,
  hist.bins = 30,
  return.tbl = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{e}{ENMevaluation object}

\item{ref.data}{Character: the reference to calculate MESS based on: occurrences ("occs") or background ("bg"), with default "occs"}

\item{sim.type}{Character: either "mess" for Multivariate Environmental Similarity Surface, "most_diff" for most different variable,
or "most_sim" for most similar variable; uses similarity function from package rmaxent}

\item{envs.var}{Character: the name of a predictor variable to plot similarities for; if left NULL, calculations are done
with respect to all variables}

\item{hist.bins}{Numeric: number of histogram bins for histogram plots; default is 30}

\item{envs}{RasterStack: environmental predictor variables used to build the models in "e"; categorical variables should be 
removed before input, as they cannot be used to calculate MESS}

\item{plot.type}{Character specifying which to plot: MESS histograms ("histogram") or MESS rasters ("raster"); default is histogram}

\item{sim.palette}{Character: the RColorBrewer palette name to use for plotting; if left NULL, the defaults are
"Set1" for discrete variables and reverse "RdYlBu" for continuous variables}

\item{palette.dir}{Numeric: direction for palette plotting; either 1 or -1}
}
\value{
A ggplot of MESS calculations for data partitions.
}
\description{
Plots environmental similarity of reference partitions (occurrences or background) 
to remaining background (all other partitions). This function does not use raster data, and thus
only calculates similarity values for data used in model training. Further, this function does
not calculate similarity for categorical variables.
}
\details{
When fully withheld testing groups are used, make sure to input either an ENMevaluation 
object or the argument occs.testing.z. In the resulting plot, partition 1 refers to the training data,
while partition 2 refers to the fully withheld testing group.

There are two variations for this plot. If "histogram", histograms are plotted showing the MESS estimates for each partition group. 
If "raster", rasters are plotted showing the geographical MESS estimates for each partition group. 
With sim.type option "mess", the similarity between environmental values associated with the 
validation occurrences (per partition group) and those associated with the entire study extent (specified by the extent 
of the input RasterStack "envs") are calculated, and the minimum similarity per grid is returned. 
Higher negative values indicate greater environmental difference between the validation occurrences
and the study extent, and higher positive values indicate greater similarity. This function uses the `rmaxent::similarity()` function 
to calculate the similarities. See the below reference for details on MESS.
}
\references{
Elith J., M. Kearney M., and S. Phillips, 2010. The art of modelling range-shifting species. Methods in Ecology and Evolution 1:330-342.
}
