% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{evalplot.respCurve.dens}
\alias{evalplot.respCurve.dens}
\title{Response curve and density plots for one variable}
\usage{
evalplot.respCurve.dens(
  mod,
  data,
  envs,
  var,
  fun = mean,
  type = c(1, 2),
  exp.curve = 0.025,
  nr.curve = 100,
  clamp.tails = TRUE,
  bw.envs = 10
)
}
\arguments{
\item{mod}{A maxent.jar or maxnet model object.}

\item{data}{Data frame of training data (occurrences + background).}

\item{envs}{Raster data (SpatRaster) of environmental variables for model projection.}

\item{var}{A character string specifying the variable name for the response curve.}

\item{fun}{A function to compute constant values for other variables (default is `median`).}

\item{type}{Number (1 or 2) to specify type of response curve to plot. See details for explanation.}

\item{exp.curve}{Numeric value indicating the range expansion for plotting (default is 0.025).}

\item{nr.curve}{Integer specifying the number of points for the response curve (default is 100).}

\item{clamp.tails}{Logical; if `TRUE`, clamping tails in plot (default is `TRUE`).}

\item{bw.envs}{The smoothing bandwidth to be used in the environmental variables}
}
\value{
A combined patchwork plot of all response curves with a shared y-axis label.
}
\description{
A wrapper function to plot response curves and density plot.
}
\examples{
\dontrun{
occs <- read.csv(file.path(system.file(package="predicts"), "/ex/bradypus.csv"))[,2:3]
envs <- rast(list.files(path=paste(system.file(package="predicts"), "/ex", sep=""), 
                        pattern="tif$", full.names=TRUE))
# No biome
envs <- envs[[!(names(envs) \%in\% "biome")]]
occs.z <- cbind(occs, terra::extract(envs, occs, ID = FALSE))
bg <- as.data.frame(predicts::backgroundSample(envs, n = 10000))
names(bg) <- names(occs)
bg.z <- cbind(bg, terra::extract(envs, bg, ID = FALSE))
os <- list(abs.auc.diff = FALSE, pred.type = "cloglog", validation.bg = "partition")
ps <- list(orientation = "lat_lat")
# Transfer envs
tr_envs <- envs * 1.5
# Plot
mod <- e@models[[1]]
# Define data as combined training values with coordinates removed
data <- rbind(e@occs, e@bg)[,3:11]
evalplot.respCurve.dens(mod, data, envs = tr_envs, var = "bio1", fun = median)
}
}
\author{
Gonzalo E. Pinilla-Buitrago
}
