% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{evalplot.respCurves}
\alias{evalplot.respCurves}
\title{Plot Response Curves for All Variables with Shared Y-Axis}
\usage{
evalplot.respCurves(
  mod,
  data,
  envs,
  fun = mean,
  type = c(1, 2),
  exp.curve = 0.025,
  nr.curve = 100,
  clamp.tails = TRUE
)
}
\arguments{
\item{mod}{A maxent.jar or maxnet model object.}

\item{data}{Data frame of training data (occurrences + background).}

\item{envs}{Raster data (SpatRaster) of environmental variables for model projection.}

\item{fun}{A function to compute constant values for other variables (default is `median`).}

\item{type}{Number (1 or 2) to specify type of response curve to plot. See details for explanation.}

\item{exp.curve}{Numeric value indicating the range expansion for plotting (default is 0.025).}

\item{nr.curve}{Integer specifying the number of points for the response curve (default is 100).}

\item{clamp.tails}{Logical; if `TRUE`, clamping tails in plot (default is `TRUE`).}
}
\value{
A combined patchwork plot of all response curves with a shared y-axis label.
}
\description{
A wrapper function to plot response curves for all contributing variables and combine
them using patchwork. The plots share a common y-axis label.
}
\examples{
\dontrun{
library(ENMeval)
occs <- read.csv(file.path(system.file(package="predicts"), "/ex/bradypus.csv"))[,2:3]
envs <- terra::rast(list.files(path=paste(system.file(package="predicts"), "/ex", sep=""), 
                        pattern="tif$", full.names=TRUE))
# No biome
envs <- envs[[!(names(envs) \%in\% "biome")]]
occs.z <- cbind(occs, terra::extract(envs, occs, ID = FALSE))
bg <- as.data.frame(predicts::backgroundSample(envs, n = 10000))
names(bg) <- names(occs)
bg.z <- cbind(bg, terra::extract(envs, bg, ID = FALSE))
os <- list(abs.auc.diff = FALSE, pred.type = "cloglog", validation.bg = "partition")
ps <- list(orientation = "lat_lat")
e <- ENMevaluate(occs, envs, bg, tune.args = list(fc = "LQ", rm = 1),
                 partitions = "block", other.settings = os, 
                 partition.settings = ps, algorithm = "maxnet", overlap = TRUE)
# Transfer envs
tr_envs <- envs * 1.5
mod <- e@models[[1]]
# Define data as combined training values with coordinates removed
data <- rbind(e@occs, e@bg)[,3:11]
# Plot
evalplot.respCurves(mod, data, envs = tr_envs)
}
}
\author{
Gonzalo E. Pinilla- Buitrago
}
