% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{loadENMevaluation}
\alias{loadENMevaluation}
\title{Load ENMevaluation object}
\usage{
loadENMevaluation(filename)
}
\arguments{
\item{filename}{character: path to the .rds file to load}
}
\description{
Load an ENMevaluation object as an .rds file. This is necessary
to use instead of \code{readRDS()} because wrapped terra SpatRasters require 
\code{unwrap()} after loading for the raster data. This convenience function does 
that for you.
}
