% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{partition.settings}
\alias{partition.settings}
\title{Partition settings}
\usage{
ps <- list(kfolds = 5)
ps <- list(orientation = "lat_lat")
ps <- list(aggregation.factor = c(4,4))
ENMevaluate(..., partition.settings = ps)
}
\arguments{
\item{orientation}{character: one of "lat_lon", "lon_lat", "lat_lat", or "lon_lon" (required for block partition)}

\item{aggregation.factor}{numeric vector: one or two numbers specifying
the factor with which to aggregate the envs raster to assign partitions (required for the checkerboard partitions)}

\item{kfolds}{Required for the random partition. This specifies the number of random partition groups, or folds, to make.}
}
\description{
This is a named list used to specify certain settings for partitioning schema.
It is inserted as an argument to ENMevaluate(). Some partitioning schema require a setting to
be specified. It is helpful to use the evalplot.grps() plotting function to visualize differences in settings.
}
\details{
For the block partition, the orientation specifications are abbreviations for "latitude" and "longitude", and 
they determine the order and orientations with which the block partitioning function creates the partition groups. For example,
"lat_lon" will split the occurrence localities first by latitude, then by longitude.

For the checkerboard partitions, the aggregation factor specifies how much to aggregate the existing cells in the envs raster
to make new spatial partitions. For example, checkerboard1 with an aggregation factor value of 2 will make the grid cells 
4 times larger and then assign occurrence and background records to partition groups based on which cell they are in. 
The checkerboard2 partition is hierarchical, so cells are first aggregated to define groups like checkerboard1, but a 
second aggregation is then made to separate the resulting 2 bins into 4 bins. For checkerboard2, two different numbers can be used
to specify the two levels of the hierarchy, or if a single number is inserted, that value will be used for both levels.
}
