% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{saveENMevaluation}
\alias{saveENMevaluation}
\title{Save ENMevaluation object}
\usage{
saveENMevaluation(e, filename)
}
\description{
Save an ENMevaluation object as an .rds file. This is necessary
to use instead of saveRDS() because terra SpatRasters require wrap() before
saving to preserve the connections to the raster data. This convenience 
function does that for you.
}
