% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune.enm.R
\name{tune.enm}
\alias{tune.enm}
\alias{tune.parallel}
\alias{tune.regular}
\alias{cv.enm}
\alias{tune.train}
\alias{tune.validate}
\title{Iterate tuning of ENMs}
\usage{
tune.train(
  enm,
  occs.z,
  bg.z,
  mod.full,
  mod.full.pred,
  envs,
  other.settings,
  partitions,
  quiet
)

tune.validate(
  enm,
  occs.train.z,
  occs.val.z,
  bg.train.z,
  bg.val.z,
  mod.k,
  nk,
  other.settings,
  partitions,
  user.eval,
  quiet
)

tune.parallel(
  d,
  envs,
  enm,
  partitions,
  tune.tbl,
  other.settings,
  partition.settings,
  user.val.grps,
  occs.testing.z,
  numCores,
  parallelType,
  user.eval,
  quiet
)

tune.regular(
  d,
  envs,
  enm,
  partitions,
  tune.tbl,
  other.settings,
  partition.settings,
  user.val.grps,
  occs.testing.z,
  updateProgress,
  user.eval,
  quiet
)

cv.enm(
  d,
  envs,
  enm,
  partitions,
  tune.i,
  other.settings,
  partition.settings,
  user.val.grps,
  occs.testing.z,
  user.eval,
  quiet
)
}
\arguments{
\item{enm}{Object of class \link{ENMdetails}.}

\item{envs}{Raster* object of environmental variables (must be in same geographic projection as occurrence data)}

\item{other.settings}{list of settings from \code{ENMevaluate()} containing other.args, pred.type, abs.auc.diff, validation.bg}

\item{partitions}{character of name of partitioning technique (see \code{?partitions})}

\item{d}{data frame from \code{ENMevaluate()} with occurrence and background coordinates (or coordinates plus predictor variable values) and partition group values}

\item{tune.tbl}{Data frame of tuning parameter combinations.}

\item{numCores}{boolean (TRUE or FALSE); if TRUE, use specifed number of cores for parallel processing}

\item{parallelType}{character of either "doParallel" or "doSNOW" to conduct parallelization}

\item{tune.i}{vector of single set of tuning parameters}
}
\description{
Internal functions to tune and summarize results for ecological niche models (ENMs) iteratively across a range of user-specified tuning settings. 
Function \code{tune.parallel()} tunes ENMs with parallelization. Function \code{cv.enm()} calculates training and validation evaluation statistics for one set of specified tuning parameters.

Validation CBI is calculated here with background values, not raster data, in order
}
